Assignments/AI-Speech-VoiceAssistant/requirements.txt
requests
python-dotenv
TTS
torch
sounddevice
soundfile